requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq==0.4.2
requests==2.31.0
gunicorn==21.2.0
sqlalchemy==2.0.23
werkzeug==3.0.1
requests